Move getDocument to io-sign package
